Workspace helper decorators are added
